# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Smart Crop Advisory System API
  version: 1.0.0
  description: >
    RESTful API for the Smart Crop Advisory System — a multilingual, AI-powered
    platform for small and marginal farmers in India.
servers:
  - url: /api/v1
    description: Version 1

# ─── Security ─────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [status, data, error]
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          nullable: true
        error:
          type: object
          nullable: true
          example: null
    ErrorEnvelope:
      type: object
      required: [status, data, error]
      properties:
        status:
          type: string
          enum: [error]
        data:
          type: object
          nullable: true
          example: null
        error:
          $ref: '#/components/schemas/ErrorObject'
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - DUPLICATE_MOBILE
            - INVALID_OTP
            - INVALID_REFRESH_TOKEN
            - INCOMPLETE_SOIL_PROFILE
            - NO_SOIL_PROFILE
            - MISSING_PARAM
            - MISSING_FILE
            - UNSUPPORTED_MEDIA_TYPE
            - PAYLOAD_TOO_LARGE
            - ADVISORY_ENGINE_UNAVAILABLE
            - VISION_ENGINE_UNAVAILABLE
            - WEATHER_UNAVAILABLE
            - MARKET_UNAVAILABLE
            - OTP_SERVICE_UNAVAILABLE
            - STT_ERROR
            - TTS_ERROR
            - INTERNAL_ERROR
        message:
          type: string
        field:
          type: string
          description: Identifies the invalid field (validation errors only)

    # ── Farmer ────────────────────────────────────────────────────────────────
    FarmerProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mobileNumber:
          type: string
          example: '+919876543210'
        name:
          type: string
          example: Gurpreet Singh
        preferredLang:
          type: string
          enum: [en, hi, pa]
        village:
          type: string
        district:
          type: string
        state:
          type: string
        landSizeAcres:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FarmerProfileUpdate:
      type: object
      properties:
        name:
          type: string
        preferredLang:
          type: string
          enum: [en, hi, pa]
        village:
          type: string
        district:
          type: string
        state:
          type: string
        landSizeAcres:
          type: number
          format: float

    # ── Soil Profile ──────────────────────────────────────────────────────────
    SoilProfileData:
      type: object
      required: [soilType, ph]
      properties:
        plotName:
          type: string
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        soilType:
          type: string
          example: loamy
        ph:
          type: number
          format: float
          minimum: 0
          maximum: 14
        nitrogen:
          type: number
          format: float
          minimum: 0
        phosphorus:
          type: number
          format: float
          minimum: 0
        potassium:
          type: number
          format: float
          minimum: 0
    SoilProfile:
      allOf:
        - type: object
          properties:
            id:
              type: string
              format: uuid
            farmerId:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
        - $ref: '#/components/schemas/SoilProfileData'

    # ── Advisory ──────────────────────────────────────────────────────────────
    CropRecommendation:
      type: object
      properties:
        cropName:
          type: string
        rank:
          type: integer
        expectedYieldRange:
          type: string
          example: '25-30 quintals/acre'
        waterRequirement:
          type: string
          example: '400-500 mm'
        estimatedInputCost:
          type: string
          example: '₹8,000/acre'
    FertilizerApplication:
      type: object
      properties:
        fertilizer:
          type: string
        quantity:
          type: string
          example: '50 kg/acre'
        timing:
          type: string
          example: 'At sowing'
        type:
          type: string
          enum: [chemical, organic]
    FertilizerSchedule:
      type: object
      properties:
        cropId:
          type: string
        cropName:
          type: string
        applications:
          type: array
          items:
            $ref: '#/components/schemas/FertilizerApplication'
        soilAmendments:
          type: array
          items:
            type: string
          description: Lime/sulfur recommendations when pH is out of optimal range

    # ── Image Analysis ────────────────────────────────────────────────────────
...
    DashboardReport:
      type: object
      properties:
        totalSessions:
          type: integer
        averageRating:
          type: number
          format: float
        sessionsByType:
          type: object
          additionalProperties:
            type: integer
        ratingDistribution:
          type: object
          additionalProperties:
            type: integer

  # ── Reusable responses ─────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            status: error
            data: null
            error:
              code: UNAUTHORIZED
              message: Authentication required
    Forbidden:
      description: Insufficient role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            status: error
            data: null
            error:
              code: FORBIDDEN
              message: Access restricted to officers and administrators
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ServiceUnavailable:
      description: Upstream service unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'

# ─── Default security (overridden per-operation for public endpoints) ──────────
security:
  - bearerAuth: []

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /auth/register:
    post:
      summary: Send OTP to mobile number
      operationId: authRegister
      tags: [Auth]
      security: []   # public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mobileNumber]
              properties:
                mobileNumber:
                  type: string
                  example: '+919876543210'
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: OTP sent
        '400':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /auth/verify-otp:
    post:
      summary: Verify OTP and receive tokens
      operationId: authVerifyOtp
      tags: [Auth]
      security: []   # public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mobileNumber, otp]
              properties:
                mobileNumber:
                  type: string
                otp:
                  type: string
                  example: '123456'
      responses:
        '200':
          description: OTP verified — tokens issued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accessToken:
                            type: string
                          refreshToken:
                            type: string
                          farmerId:
                            type: string
                            format: uuid
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      summary: Refresh access token
      operationId: authRefresh
      tags: [Auth]
      security: []   # public (uses refresh token in body)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [farmerId, refreshToken]
              properties:
                farmerId:
                  type: string
                  format: uuid
                refreshToken:
                  type: string
      responses:
        '200':
          description: New access and refresh tokens issued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accessToken:
                            type: string
                          refreshToken:
                            type: string
                          farmerId:
                            type: string
                            format: uuid
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      summary: Logout and invalidate refresh token
      operationId: authLogout
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: Logged out successfully
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Farmers ─────────────────────────────────────────────────────────────────
  /farmers/me:
    get:
      summary: Get authenticated farmer's profile
      operationId: getFarmerMe
      tags: [Farmers]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Farmer profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FarmerProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update authenticated farmer's profile
      operationId: updateFarmerMe
      tags: [Farmers]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FarmerProfileUpdate'
      responses:
        '200':
          description: Updated farmer profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FarmerProfile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Soil Profiles ────────────────────────────────────────────────────────────
  /soil-profiles:
    post:
      summary: Create a new soil profile
      operationId: createSoilProfile
      tags: [Soil Profiles]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SoilProfileData'
      responses:
        '201':
          description: Soil profile created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SoilProfile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /soil-profiles/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: Get a soil profile by ID
      operationId: getSoilProfile
      tags: [Soil Profiles]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Soil profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SoilProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update a soil profile
      operationId: updateSoilProfile
      tags: [Soil Profiles]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SoilProfileData'
      responses:
        '200':
          description: Updated soil profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SoilProfile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Advisory ─────────────────────────────────────────────────────────────────
  /advisory/crops:
    get:
      summary: Get crop recommendations for a plot
      operationId: getCropRecommendations
      tags: [Advisory]
      security:
        - bearerAuth: []
      parameters:
        - name: plotId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ranked crop recommendations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/CropRecommendation'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /advisory/fertilizer:
    get:
      summary: Get fertilizer schedule for a crop and plot
      operationId: getFertilizerGuidance
      tags: [Advisory]
      security:
        - bearerAuth: []
      parameters:
        - name: plotId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: cropId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fertilizer schedule
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FertilizerSchedule'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  # ── Images ───────────────────────────────────────────────────────────────────
...
  /feedback:
    post:
      summary: Submit or dismiss feedback for an advisory session
      operationId: submitFeedback
      tags: [Feedback]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId]
              properties:
                sessionId:
                  type: string
                  format: uuid
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  description: Rating 1–5 (provide either rating or dismiss)
                dismiss:
                  type: boolean
                  description: Set true to dismiss without rating
      responses:
        '200':
          description: Feedback recorded
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ok:
                            type: boolean
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Dashboard ─────────────────────────────────────────────────────────────────
  /dashboard/reports:
    get:
      summary: Get aggregated usage and feedback reports (officer/admin only)
      operationId: getDashboardReports
      tags: [Dashboard]
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          required: true
          schema:
            type: string
            enum: [officer, admin]
          description: Caller's role (must be officer or admin)
      responses:
        '200':
          description: Aggregated report
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DashboardReport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ── Notifications ─────────────────────────────────────────────────────────────
...